+++ native/wasmex/src/module.rs
@@ -1,7 +1,7 @@
 use rustler::{resource::ResourceArc, types::binary::Binary, NifResult};
 use std::sync::Mutex;
 
-use wasmer::{Module, Store};
+use wasmer::{wat2wasm, Module, Store};
 
 use crate::atoms;
 
@@ -18,8 +18,14 @@ pub struct ModuleResourceResponse {
 #[rustler::nif(name = "module_compile")]
 pub fn compile(binary: Binary) -> NifResult<ModuleResourceResponse> {
     let bytes = binary.as_slice();
+    let bytes = wat2wasm(bytes).map_err(|e| {
+        rustler::Error::Term(Box::new(format!(
+            "Error while parsing bytes: {}.",
+            e.to_string()
+        )))
+    })?;
     let store = Store::default();
-    match Module::new(&store, &bytes) {
+    match Module::new(&store, bytes) {
         Ok(module) => {
             let resource = ResourceArc::new(ModuleResource {
                 module: Mutex::new(module),