+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Types of changes
 
 * `Wasmex.Module.compile/1` which can take the bytes of a .wasm files and compiles the WASM module. This module can be given to the new methods `Wasmex.Instance.new/2` and `Wasmex.Instance.new_wasi/3` allowing to re-used precompiled modules. This has a big potential speed-up if one wishes to run a WASI instance multiple times. For example, the wasmex test suite went from 14.5s to just 0.2s runtime.
 * `Wasmex.start_link` can now also be called with a precompiled module in addition to the .wasm file bytes.
+* `Wasmex.Module.compile/1` can now parse WebAssembly text format (WAT) too
+* WASM modules without exported memory can now be instantiated
 
 ### Deprecated